splitted checks to two files
