+++ dfirtrack_main/importer/file/csv_import.py
@@ -3,7 +3,7 @@ from django.contrib import messages
 from django.utils import timezone
 from dfirtrack_config.models import SystemImporterFileCsvConfigModel
 from dfirtrack_main.importer.file.csv_add_attributes import add_fk_attributes, add_many2many_attributes, create_lock_tags
-from dfirtrack_main.importer.file.csv_check_data import run_check_config_cron_user, run_check_content_file_system, run_check_content_file_type
+from dfirtrack_main.importer.file.csv_run_checks import run_check_config_cron_user, run_check_content_file_system, run_check_content_file_type
 from dfirtrack_main.importer.file.csv_messages import final_messages
 from dfirtrack_main.logger.default_logger import info_logger, warning_logger
 from dfirtrack_main.models import System
